refactor: styles, types, storybook config
