+++ .storybook/main.ts
@@ -3,11 +3,11 @@ import type { StorybookConfig } from '@storybook/react-vite';
 const config: StorybookConfig = {
 	stories: ['../src/**/*.mdx', '../src/**/*.stories.@(js|jsx|mjs|ts|tsx)'],
 	addons: [
-		'@storybook/addon-onboarding',
 		'@storybook/addon-essentials',
 		'@chromatic-com/storybook',
 		'@storybook/addon-interactions',
 		'@storybook/addon-themes',
+		'@storybook/addon-docs',
 	],
 	framework: {
 		name: '@storybook/react-vite',
