+++ .storybook/storybook-global.css
@@ -1,0 +1,13 @@
+#storybook-docs {
+	overflow: auto;
+	height: 100vh;
+}
+
+.story_container {
+	display: flex;
+	align-items: center;
+	justify-content: center;
+
+	width: 500px;
+	height: 500px;
+}
