+++ src/components/not-found/not-found.module.css
@@ -8,7 +8,7 @@
 .content {
 	display: flex;
 	flex-direction: column;
-	gap: var(--gap-l);
+	gap: var(--gap-lg);
 	align-items: center;
 }
