+++ src/modules/dashboard/components/script-template-list/script-template-item/script-template-item.skeleton.tsx
@@ -1,5 +1,5 @@
 import Skeleton from 'react-loading-skeleton';
 
 export const ScriptTemplateItemSkeleton = () => (
-	<Skeleton style={{ padding: 'var(--gap-l)' }} />
+	<Skeleton style={{ padding: 'var(--gap-lg)' }} />
 );
