+++ src/utils/types/direction.ts
@@ -1,0 +1,1 @@
+export type Direction = 'horizontal' | 'vertical';
