+++ src/utils/types/index.ts
@@ -1,1 +1,2 @@
-export * from './filter.types';
+export * from './filter';
+export * from './direction';
